refactor(zim): extract _SearchMixin to zim/search.py
Move search/suggestion/title-resolution methods (search_zim_file,
search_with_filters, get_search_suggestions, find_entry_by_title,
search_all, _find_entry_by_search and helpers) into _SearchMixin.
ZimOperations now inherits from _SearchMixin.

The mixin reaches back to openzim_mcp.zim_operations for late-bound
lookups of zim_archive, Query, Searcher, SuggestionSearcher, and
PaginationCursor so existing test patches against those symbols
continue to work without changes.
