+++ openzim_mcp/zim/__init__.py
@@ -1,21 +1,14 @@
 """ZIM operations package.
 
-Public API surface is preserved by re-exporting ``ZimOperations`` and the
-module-level helpers/constants here. External callers should keep importing
-from ``openzim_mcp.zim_operations`` (the backward-compat shim) — this
-package's internal layout is an implementation detail.
+The package layout is an implementation detail; external callers should
+import from ``openzim_mcp.zim_operations`` (the backward-compat shim).
+
+This ``__init__`` deliberately defers the public re-exports until late
+to avoid circularity during the in-progress refactor: ``zim_operations``
+still owns the ``ZimOperations`` class and ``zim_archive`` symbol, and
+each mixin module under this package needs to import
+``openzim_mcp.zim_operations`` for late-bound access. Re-exporting
+``ZimOperations`` from this package's ``__init__`` would force
+``zim_operations`` to import this package before ``ZimOperations`` is
+defined.
 """
-
-from openzim_mcp.zim.archive import (  # noqa: F401
-    MAX_REDIRECT_DEPTH,
-    PaginationCursor,
-    ZimOperations,
-    zim_archive,
-)
-
-__all__ = [
-    "MAX_REDIRECT_DEPTH",
-    "PaginationCursor",
-    "ZimOperations",
-    "zim_archive",
-]
